Readability Problem Description
Readability
First program Solution for CS50 2019
problem Description
Implement a program that computes the approximate grade level needed to comprehend some text, per the below.
$ ./readability Text: Congratulations! Today is your day. You're off to Great Places! You're off and away! Grade 3

One such readability test is the Coleman-Liau index. The Coleman-Liau index of a text is designed to output what (U.S.) grade level is needed to understand the text. The formula is:

index = 0.0588 * L - 0.296 * S - 15.8

Here, L is the average number of letters per 100 words in the text, and S is the average number of sentences per 100 words in the text.

Let’s write a program called readability that takes a text and determines its reading level.